requirements.txt
beautifulsoup4==4.11.1
ImageHash==4.3.1
Js2Py==0.74
Pillow==9.3.0
pillow_avif_plugin==1.3.1
requests==2.28.1